requirements.txt
openvino-dev[pytorch,onnx]==2022.3.0
onnxruntime-gpu==1.14.1
onnxconverter-common==1.13.0
pycuda==2022.2.2
openpyxl==3.1.2